#180 Refactor: change value of param, split method create header element to a file component
