+++ javascript/big-practice/src/scripts/views/menuView.js
@@ -38,7 +38,7 @@ export default class MenuView {
           this.elementHelpers.addClass(menu[sessionStorage.getItem(STORAGE_KEYS.PAGE_NUMBER)], 'menu-color');
 
           renderTabs();
-          changeLogoFollowTab(e.target.querySelector('h1').textContent);
+          changeLogoFollowTab(e.target.querySelector('span').textContent);
         } else {
           this.renderPopupError('Page number is not found');
         }
